Fix test sktime forecaster tests for y univariate, multivariate and both (#1306)
* includes condtion for multivariate and removes assert for Pd.Series

* includes condtion for multivariate

* includes condition for multivariate y

* includes condition for both  y

* includes condition for both  y

* refactors condition for both, univariate and mulitvariate

* include get_n_columns in the utils
