+++ sktime/forecasting/tests/test_sktime_forecasters.py
@@ -3,7 +3,7 @@
 # copyright: sktime developers, BSD-3-Clause License (see LICENSE file)
 
 # test API provided through BaseSktimeForecaster
-
+"""SKtime Forecasters Test."""
 __author__ = ["@mloning"]
 __all__ = [
     "test_different_fh_in_fit_and_predict_req",
@@ -18,7 +18,6 @@ __all__ = [
 ]
 
 import numpy as np
-import pandas as pd
 import pytest
 
 from sktime.forecasting.base import BaseForecaster
@@ -27,6 +26,8 @@ from sktime.forecasting.model_selection import temporal_train_test_split
 from sktime.registry import all_estimators
 from sktime.utils._testing.estimator_checks import _construct_instance
 from sktime.utils._testing.forecasting import make_forecasting_problem
+from sktime.utils._testing.series import _make_series
+from sktime.utils._testing.forecasting import _get_n_columns
 
 # get all forecasters
 FORECASTERS = [
@@ -42,6 +43,7 @@ WINDOW_FORECASTERS = [
     if issubclass(forecaster, _BaseWindowForecaster)
 ]
 
+
 # testing data
 y = make_forecasting_problem()
 y_train, y_test = temporal_train_test_split(y, train_size=0.75)
@@ -50,24 +52,36 @@ y_train, y_test = temporal_train_test_split(y, train_size=0.75)
 # test _y setting
 @pytest.mark.parametrize("Forecaster", FORECASTERS)
 def test_oh_setting(Forecaster):
+    """Check cuttoff and _y."""
     # check _y and cutoff is None after construction
     f = _construct_instance(Forecaster)
-    assert f._y is None
-    assert f.cutoff is None
+    n_columns_list = _get_n_columns(f.get_tag("y:scitype"))
 
-    # check that _y and cutoff is updated during fit
-    f.fit(y_train, fh=FH0)
-    assert isinstance(f._y, pd.Series)
-    assert len(f._y) > 0
-    assert f.cutoff == y_train.index[-1]
+    for n_columns in n_columns_list:
+        y = _make_series(n_columns=n_columns)
+        y_train, y_test = temporal_train_test_split(y, train_size=0.75)
 
-    # check data pointers
-    np.testing.assert_array_equal(f._y.index, y_train.index)
+        assert f._y is None
+        assert f.cutoff is None
 
-    # check that _y and cutoff is updated during update
-    f.update(y_test, update_params=False)
-    np.testing.assert_array_equal(f._y.index, np.append(y_train.index, y_test.index))
-    assert f.cutoff == y_test.index[-1]
+        # check that _y and cutoff is updated during fit
+        f.fit(y_train, fh=FH0)
+        # assert isinstance(f._y, pd.Series)
+        # action:uncomments the line above
+        # why: fails for multivariates cause they are DataFrames
+        # solution: look for a general solution for Series and DataFrames
+        assert len(f._y) > 0
+        assert f.cutoff == y_train.index[-1]
+
+        # check data pointers
+        np.testing.assert_array_equal(f._y.index, y_train.index)
+
+        # check that _y and cutoff is updated during update
+        f.update(y_test, update_params=False)
+        np.testing.assert_array_equal(
+            f._y.index, np.append(y_train.index, y_test.index)
+        )
+        assert f.cutoff == y_test.index[-1]
 
 
 # check setting/getting API for forecasting horizon
@@ -85,6 +99,7 @@ FORECASTERS_OPTIONAL = [
 # testing Forecasters which require fh during fitting
 @pytest.mark.parametrize("Forecaster", FORECASTERS_REQUIRED)
 def test_no_fh_in_fit_req(Forecaster):
+    """Check if fh is required in fit."""
     f = _construct_instance(Forecaster)
     # fh required in fit, raises error if not passed
     with pytest.raises(ValueError):
@@ -93,6 +108,7 @@ def test_no_fh_in_fit_req(Forecaster):
 
 @pytest.mark.parametrize("Forecaster", FORECASTERS_REQUIRED)
 def test_fh_in_fit_req(Forecaster):
+    """Checks if fh is requred in fit."""
     f = _construct_instance(Forecaster)
     f.fit(y_train, fh=FH0)
     np.testing.assert_array_equal(f.fh, FH0)
@@ -102,6 +118,7 @@ def test_fh_in_fit_req(Forecaster):
 
 @pytest.mark.parametrize("Forecaster", FORECASTERS_REQUIRED)
 def test_same_fh_in_fit_and_predict_req(Forecaster):
+    """Check if fh is the same in fit and predict."""
     f = _construct_instance(Forecaster)
     f.fit(y_train, fh=FH0)
     np.testing.assert_array_equal(f.fh, FH0)
@@ -111,6 +128,7 @@ def test_same_fh_in_fit_and_predict_req(Forecaster):
 
 @pytest.mark.parametrize("Forecaster", FORECASTERS_REQUIRED)
 def test_different_fh_in_fit_and_predict_req(Forecaster):
+    """Check if fh is different in fit and predict."""
     f = _construct_instance(Forecaster)
     f.fit(y_train, fh=FH0)
     np.testing.assert_array_equal(f.fh, FH0)
@@ -123,37 +141,62 @@ def test_different_fh_in_fit_and_predict_req(Forecaster):
 # testing Forecasters which take fh either during fitting or predicting
 @pytest.mark.parametrize("Forecaster", FORECASTERS_OPTIONAL)
 def test_no_fh_opt(Forecaster):
+    """Check if fh is optional in fit."""
     f = _construct_instance(Forecaster)
-    f.fit(y_train)
-    # not passing fh to either fit or predict raises error
-    with pytest.raises(ValueError):
-        f.predict()
+    n_columns_list = _get_n_columns(f.get_tag("y:scitype"))
+
+    for n_columns in n_columns_list:
+        y_train = _make_series(n_columns=n_columns)
+        f.fit(y_train)
+        # not passing fh to either fit or predict raises error
+        with pytest.raises(ValueError):
+            f.predict()
 
 
 @pytest.mark.parametrize("Forecaster", FORECASTERS_OPTIONAL)
 def test_fh_in_fit_opt(Forecaster):
+    """Check if fh is optional in fit."""
     f = _construct_instance(Forecaster)
-    f.fit(y_train, fh=FH0)
-    np.testing.assert_array_equal(f.fh, FH0)
-    f.predict()
-    np.testing.assert_array_equal(f.fh, FH0)
+    n_columns_list = _get_n_columns(f.get_tag("y:scitype"))
+
+    for n_columns in n_columns_list:
+        y_train = _make_series(n_columns=n_columns)
+        f = _construct_instance(Forecaster)
+        f.fit(y_train, fh=FH0)
+        np.testing.assert_array_equal(f.fh, FH0)
+        f.predict()
+        np.testing.assert_array_equal(f.fh, FH0)
 
 
 @pytest.mark.parametrize("Forecaster", FORECASTERS_OPTIONAL)
 def test_fh_in_predict_opt(Forecaster):
+    """Check if fh is optional in predict."""
     f = _construct_instance(Forecaster)
-    f.fit(y_train)
-    f.predict(FH0)
-    np.testing.assert_array_equal(f.fh, FH0)
+    n_columns_list = _get_n_columns(f.get_tag("y:scitype"))
+
+    for n_columns in n_columns_list:
+        y_train = _make_series(n_columns=n_columns)
+        y_train = _make_series(n_columns=1)
+        f = _construct_instance(Forecaster)
+        f.fit(y_train)
+        f.predict(FH0)
+        np.testing.assert_array_equal(f.fh, FH0)
 
 
 @pytest.mark.parametrize("Forecaster", FORECASTERS_OPTIONAL)
 def test_same_fh_in_fit_and_predict_opt(Forecaster):
+    """Check if fh is the same in fit and predict."""
     f = _construct_instance(Forecaster)
-    # passing the same fh to both fit and predict works
-    f.fit(y_train, fh=FH0)
-    f.predict(FH0)
-    np.testing.assert_array_equal(f.fh, FH0)
+    n_columns_list = _get_n_columns(f.get_tag("y:scitype"))
+
+    for n_columns in n_columns_list:
+        y_train = _make_series(n_columns=n_columns)
+
+        # passing the same fh to both fit and predict works
+
+        f.fit(y_train, fh=FH0)
+        f.predict(FH0)
+        np.testing.assert_array_equal(f.fh, FH0)
 
 
 @pytest.mark.parametrize("Forecaster", WINDOW_FORECASTERS)
